requirements.txt
streamlit
mysql-connector-python
st-pages
streamlit_extras
numpy
influxdb-client
influxdb3-python
streamlit_javascript
scipy
